requirements.txt
fastapi
uvicorn
tensorflow
pillow
opencv-python
numpy